+++ backend/core/src/seeder/seeder.module.ts
@@ -24,6 +24,7 @@ import { ListingDefaultOnePreferenceSeed } from "../seeds/listings/listing-defau
 import { ListingDefaultNoPreferenceSeed } from "../seeds/listings/listing-default-no-preference-seed"
 import { Preference } from "../preferences/entities/preference.entity"
 import { ListingDefaultFCFSSeed } from "../seeds/listings/listing-default-fcfs-seed"
+import { ListingDefaultOpenSoonSeed } from "../seeds/listings/listing-default-open-soon"
 import { ListingTritonSeed } from "../seeds/listings/listing-triton-seed"
 import { ListingDefaultBmrChartSeed } from "../seeds/listings/listing-default-bmr-chart-seed"
 import { ApplicationMethod } from "../application-methods/entities/application-method.entity"
@@ -76,6 +77,8 @@ export class SeederModule {
         ListingColiseumSeed,
         ListingDefaultOnePreferenceSeed,
         ListingDefaultNoPreferenceSeed,
+        ListingDefaultFCFSSeed,
+        ListingDefaultOpenSoonSeed,
         ListingDefaultBmrChartSeed,
         ListingTritonSeed,
         ListingDefaultReservedSeed,