+++ ui-components/src/page_components/listing/ListingsList.tsx
@@ -42,17 +42,8 @@ const ListingsList = (props: ListingsProps) => {
     if (openDateState(listing)) {
       const date = listing.applicationOpenDate
       const openDate = moment(date)
-      formattedDate = openDate.format("MMMM D, YYYY")
-
-      // XXX: the line below is logic that was previously in /ui-components/src/notifications/ApplicationStatus.tsx
-      // `vivid` is a prop on ApplicationStatus that determines background color and text size
-      // but it was also used to determine content
-      // what is the corresponding date to determine content here?
-      //
-      // content = vivid ? t("listings.comingSoon") : t("listings.applicationOpenPeriod")
-      //
-      // for now, use listings.comingSoon
-      content = t("listings.comingSoon")
+      formattedDate = openDate.format("MMM. D, YYYY")
+      content = t("listings.applicationOpenPeriod")
     } else {
       if (listing.applicationDueDate) {
         const dueDate = moment(listing.applicationDueDate)